lesson35/visual_regression_test/requirements.txt
playwright>=1.40.0
pillow>=10.0.0
pytest>=7.4.0
pytest-playwright>=0.4.0